+++ src/views/AppNavigation.vue
@@ -88,7 +88,7 @@ License along with this library.  If not, see <http://www.gnu.org/licenses/>.
 			</AppNavigationItem>
 		</template>
 		<template #footer>
-			<TheSettings />
+			<AppNavigationSettings />
 		</template>
 	</AppNavigation>
 </template>
@@ -96,7 +96,7 @@ License along with this library.  If not, see <http://www.gnu.org/licenses/>.
 <script>
 import ListItemCalendar from '../components/AppNavigation/ListItemCalendar.vue'
 import Colorpicker from '../components/AppNavigation/Colorpicker.vue'
-import TheSettings from '../components/AppNavigation/TheSettings.vue'
+import AppNavigationSettings from '../components/AppNavigation/AppNavigationSettings.vue'
 
 import AppNavigation from '@nextcloud/vue/dist/Components/AppNavigation'
 import AppNavigationCounter from '@nextcloud/vue/dist/Components/AppNavigationCounter'
@@ -113,7 +113,7 @@ export default {
 		AppNavigation,
 		AppNavigationItem,
 		AppNavigationCounter,
-		TheSettings,
+		AppNavigationSettings,
 		draggable,
 	},
 	directives: {
